# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PMS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoice": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Return all Invoices",
                "operationId": "InvoiceAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Create new Invoice",
                "operationId": "createInvoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paid": {
                                        "description": "Paid amount",
                                        "type": "integer",
                                        "nullable": "false"
                                    },
                                    "date": {
                                        "description": "Date of invoice created",
                                        "type": "string",
                                        "format": "date",
                                        "nullable": "false"
                                    },
                                    "customer": {
                                        "description": "Customer related to invoice",
                                        "type": "integer",
                                        "nullable": "false"
                                    },
                                    "items": {
                                        "description": "Invoice items",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "drug": {
                                                    "description": "Drug related to invoice",
                                                    "type": "integer",
                                                    "nullable": "false"
                                                },
                                                "qty": {
                                                    "description": "Quantity of drug",
                                                    "type": "integer",
                                                    "nullable": "false"
                                                },
                                                "cost": {
                                                    "description": "Cost of drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "price": {
                                                    "description": "Price of drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "discount": {
                                                    "description": "Amount of discount for drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "exp": {
                                                    "description": "Expiration date of drug",
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": "false"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invoice/{id}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Find Invoice by id",
                "operationId": "getInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Edit Invoice by id",
                "operationId": "editInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Invoice identifier",
                                        "type": "integer",
                                        "nullable": "false"
                                    },
                                    "paid": {
                                        "description": "Paid amount",
                                        "type": "integer",
                                        "nullable": "false"
                                    },
                                    "date": {
                                        "description": "Date of invoice created",
                                        "type": "string",
                                        "format": "date",
                                        "nullable": "false"
                                    },
                                    "customer": {
                                        "description": "Customer related to invoice",
                                        "type": "integer",
                                        "nullable": "false"
                                    },
                                    "items": {
                                        "description": "Invoice items",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "drug": {
                                                    "description": "Drug related to invoice",
                                                    "type": "integer",
                                                    "nullable": "false"
                                                },
                                                "qty": {
                                                    "description": "Quantity of drug",
                                                    "type": "integer",
                                                    "nullable": "false"
                                                },
                                                "cost": {
                                                    "description": "Cost of drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "price": {
                                                    "description": "Price of drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "discount": {
                                                    "description": "Amount of discount for drug",
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": "false"
                                                },
                                                "exp": {
                                                    "description": "Expiration date of drug",
                                                    "type": "string",
                                                    "format": "date",
                                                    "nullable": "false"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Delete Invoice by id",
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "402": {
                        "description": "Invoice not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomerSchema": {
                "properties": {
                    "id": {
                        "description": "Customer identifier",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "Cutomer name",
                        "type": "string",
                        "nullable": "false"
                    },
                    "contact": {
                        "description": "Customer phone number",
                        "type": "string",
                        "nullable": "false"
                    },
                    "address": {
                        "description": "Customer address",
                        "type": "string",
                        "nullable": "false"
                    },
                    "pharmacy": {
                        "$ref": "#/components/schemas/PharmacySchema"
                    }
                },
                "type": "object"
            },
...
            "InvoiceSchema": {
                "properties": {
                    "id": {
                        "description": "Invoice identifier",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "total": {
                        "description": "Invoice total price",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "totalDiscount": {
                        "description": "Invoice total discount",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "totalNet": {
                        "description": "Invoice total net",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "totalProfit": {
                        "description": "Invoice total Profit",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "paid": {
                        "description": "Paid amount",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "date": {
                        "description": "Date of invoice created",
                        "type": "string",
                        "format": "date",
                        "nullable": "false"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerSchema"
                    }
                },
                "type": "object"
            },
...
            "PharmacySchema": {
                "properties": {
                    "id": {
                        "description": "Pharmacy identifier",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "Pharmacy name",
                        "type": "string",
                        "nullable": "false"
                    },
                    "licence": {
                        "description": "Pharmacy licence id",
                        "type": "string",
                        "nullable": "false"
                    },
                    "safe": {
                        "description": "Safe amount",
                        "type": "number",
                        "format": "double",
                        "nullable": "false"
                    },
                    "chest": {
                        "description": "Chest amount",
                        "type": "number",
                        "format": "double",
                        "nullable": "false"
                    }
                },
                "type": "object"
            }
        },
...
    }
}